+++ src/Core/Contract/SetStateViewInstanceHandlerInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ApplicationManagerTools\AmDriver\Core\Contract;
+
+use ApplicationManagerTools\AmDriver\Core\Dto\OrchestrationCommand;
+
+interface SetStateViewInstanceHandlerInterface
+{
+    public function handle(OrchestrationCommand $command): void;
+}